fix(Rendering): wip on most of graphic issue in 2D
