+++ include/M3L/Rendering/Color.hpp
@@ -4,7 +4,7 @@
 
 #include "Tool/Export.hpp"
 
-#define CLR(_clr)           (static_cast<uint32_t>(_clr.A) << 24) | (static_cast<uint32_t>(_clr.B) << 16) | (static_cast<uint32_t>(_clr.G) << 8) | _clr.R
+#define CLR(_clr)           (static_cast<uint32_t>(_clr.A) << 24) | (static_cast<uint32_t>(_clr.R) << 16) | (static_cast<uint32_t>(_clr.G) << 8) | _clr.B
 #define CLR_GET_ALPHA(_clr) ((_clr >> 24) & 0xFF)
 #define CLR_GET_RGB(_clr)   (_clr & 0xFFFFFF)
