+++ src/Rendering/Sprite.cpp
@@ -38,7 +38,7 @@ namespace m3l
     {
         m_vertex.clear();
         m_vertex.append({ { 0, 0 }, m_rect.pos });
-        m_vertex.append({ { m_rect.size.x, 0 }, { m_rect.pos.x + m_rect.size.x, m_rect.pos.y }  });
+        m_vertex.append({ { m_rect.size.x, 0 }, { m_rect.pos.x + m_rect.size.x, m_rect.pos.y } });
         m_vertex.append({ { 0, m_rect.size.y }, { m_rect.pos.x, m_rect.pos.y + m_rect.size.y } });
         m_vertex.append({ m_rect.size, m_rect.pos + m_rect.size });
     }
